text_summarizer/requirement.txt
tk
python-docx
sumy
nltk